feat: rename AccessControl.types.ts to types.ts
